created a field for mission
Add the following to the mission: The University of Iowa Youth Ballet
and School of Dance strives to serve the people of Iowa by providing
high-quality outreach that enriches lives by encouraging and nuturing
artistic expression and fostering the love and teaching of movement. We
are committed to providing excellent, technique appropriate, and
challenging classes for all ability levels across dance disciplines. We
are dedicated to developing strong, well-rounded dancers and teachers
of dance.